+++ packages/office-ui-fabric-react/etc/office-ui-fabric-react.api.ts
@@ -411,7 +411,7 @@ class BasePickerListBelow<T, P extends IBasePickerProps<T>> extends BasePicker<T
 }
 
 // @public (undocumented)
-class BaseSelectedItemsList<T, P extends IBaseSelectedItemsListProps<T>> extends BaseComponent<P, IBaseSelectedItemsListState>, implements IBaseSelectedItemsList<T> {
+class BaseSelectedItemsList<T, P extends IBaseSelectedItemsListProps<T>> extends BaseComponent<P, IBaseSelectedItemsListState<T>>, implements IBaseSelectedItemsList<T> {
   constructor(basePickerProps: P);
   // (undocumented)
   addItems: (items: T[]) => void;
@@ -438,7 +438,7 @@ class BaseSelectedItemsList<T, P extends IBaseSelectedItemsListProps<T>> extends
   // (undocumented)
   protected onSelectionChanged: () => void;
   // (undocumented)
-  removeItem: (item: ISelectedItemProps<T>) => void;
+  removeItem: (item: T) => void;
   // (undocumented)
   removeItemAt: (index: number) => void;
   // (undocumented)
@@ -449,6 +449,8 @@ class BaseSelectedItemsList<T, P extends IBaseSelectedItemsListProps<T>> extends
   render(): any;
   // (undocumented)
   protected renderItems: () => JSX.Element[];
+  // (undocumented)
+  replaceItem: (itemToReplace: T, itemsToReplaceWith: T[]) => void;
   // (undocumented)
   protected root: HTMLElement;
   // (undocumented)
@@ -706,7 +708,7 @@ class CompactPeoplePickerBase extends BasePeoplePicker {
   static defaultProps: {
     createGenericItem: typeof createGenericItem;
     onRenderItem: (props: IPeoplePickerItemSelectedProps) => JSX.Element;
-    onRenderSuggestionsItem: (personaProps: IPersonaProps, suggestionsProps?: IBasePickerSuggestionsProps | undefined) => JSX.Element;
+    onRenderSuggestionsItem: (personaProps: IPersonaProps, suggestionsProps?: IBasePickerSuggestionsProps<any> | undefined) => JSX.Element;
   }
 }
 
@@ -777,6 +779,12 @@ export function correctRGB(color: IRGB): IRGB;
 
 // @public
 export function createArray<T>(size: number, getItem: (index: number) => T): T[];
+
+// @public
+export function createComponent<TComponentProps, TTokens, TStyleSet extends IStyleSet<TStyleSet>, TViewProps = TComponentProps, TStatics = {}>(component: IComponent<TComponentProps, TTokens, TStyleSet, TViewProps, TStatics>): React.StatelessComponent<TComponentProps> & TStatics;
+
+// @public
+export function createFactory<TProps>(ComponentType: React.ComponentType<TProps>, options?: IFactoryOptions<TProps>): ISlotFactory<TProps>;
 
 // @public (undocumented)
 export function createFontStyles(localeCode: string | null): IFontStyles;
@@ -1448,6 +1456,9 @@ export function getScrollbarWidth(): number;
 // @public
 export function getShade(color: IColor, shade: Shade, isInverted?: boolean): IColor | null;
 
+// @public
+export function getSlots<TProps extends TSlots, TSlots extends ISlotProps<TProps, TSlots>>(userProps: TProps, slots: ISlotDefinition<Required<TSlots>>): ISlots<Required<TSlots>>;
+
 // @public (undocumented)
 export function getSubmenuItems(item: IContextualMenuItem): IContextualMenuItem[] | undefined;
 
@@ -1956,25 +1967,8 @@ interface IBasePickerStyles {
   text: IStyle;
 }
 
-// @public (undocumented)
-interface IBasePickerSuggestionsProps {
-  className?: string;
-  forceResolveText?: string;
-  loadingText?: string;
-  mostRecentlyUsedHeaderText?: string;
-  noResultsFoundText?: string;
-  onRenderNoResultFound?: IRenderFunction<void>;
-  resultsFooter?: () => JSX.Element;
-  resultsFooterFull?: () => JSX.Element;
-  resultsMaximumNumber?: number;
-  searchForMoreText?: string;
-  searchingText?: string;
-  showRemoveButtons?: boolean;
-  suggestionsAvailableAlertText?: string;
-  suggestionsClassName?: string;
-  suggestionsContainerAriaLabel?: string;
-  suggestionsHeaderText?: string;
-  suggestionsItemClassName?: string;
+// @public
+interface IBasePickerSuggestionsProps<T = any> extends Pick<ISuggestionsProps<T>, 'onRenderNoResultFound' | 'suggestionsHeaderText' | 'mostRecentlyUsedHeaderText' | 'noResultsFoundText' | 'className' | 'suggestionsClassName' | 'suggestionsItemClassName' | 'searchForMoreText' | 'forceResolveText' | 'loadingText' | 'searchingText' | 'resultsFooterFull' | 'resultsFooter' | 'resultsMaximumNumber' | 'showRemoveButtons' | 'suggestionsAvailableAlertText' | 'suggestionsContainerAriaLabel'> {
 }
 
 // @public
@@ -2010,9 +2004,9 @@ interface IBaseSelectedItemsListProps<T> extends React.ClassAttributes<any> {
 }
 
 // @public (undocumented)
-interface IBaseSelectedItemsListState {
+interface IBaseSelectedItemsListState<T = any> {
   // (undocumented)
-  items?: any;
+  items: T[];
 }
 
 // @public (undocumented)
@@ -3066,8 +3060,20 @@ interface ICommandBarStyles {
   secondarySet?: IStyle;
 }
 
+// @public
+interface IComponent<TComponentProps, TTokens, TStyleSet extends IStyleSet<TStyleSet>, TViewProps = TComponentProps, TStatics = {}> {
+  displayName: string;
+  factoryOptions?: IFactoryOptions<TComponentProps>;
+  fields?: string[];
+  state?: IStateComponentType<TComponentProps, TViewProps>;
+  statics?: TStatics;
+  styles?: IStylesFunctionOrObject<TViewProps, TTokens, TStyleSet>;
+  tokens?: ITokenFunctionOrObject<TViewProps, TTokens>;
+  view: IViewComponent<TViewProps>;
+}
+
 // @public (undocumented)
-class IconBase extends BaseComponent<IIconProps, IIconState> {
+class IconBase extends React.PureComponent<IIconProps, IIconState> {
   constructor(props: IIconProps);
   // (undocumented)
   render(): JSX.Element;
@@ -3109,6 +3115,8 @@ enum IconNames {
   // (undocumented)
   AccessLogoFill = "AccessLogoFill",
   // (undocumented)
+  AccountBrowser = "AccountBrowser",
+  // (undocumented)
   AccountManagement = "AccountManagement",
   // (undocumented)
   Accounts = "Accounts",
@@ -3133,6 +3141,8 @@ enum IconNames {
   // (undocumented)
   AddGroup = "AddGroup",
   // (undocumented)
+  AddHome = "AddHome",
+  // (undocumented)
   AddIn = "AddIn",
   // (undocumented)
   AddNotes = "AddNotes",
@@ -3140,6 +3150,8 @@ enum IconNames {
   AddOnlineMeeting = "AddOnlineMeeting",
   // (undocumented)
   AddPhone = "AddPhone",
+  // (undocumented)
+  AddReaction = "AddReaction",
   // (undocumented)
   AddTo = "AddTo",
   // (undocumented)
@@ -3227,6 +3239,14 @@ enum IconNames {
   // (undocumented)
   AppIconDefaultList = "AppIconDefaultList",
   // (undocumented)
+  AppleTVMenu = "AppleTVMenu",
+  // (undocumented)
+  AppleTVMicrophone = "AppleTVMicrophone",
+  // (undocumented)
+  AppleTVMonitor = "AppleTVMonitor",
+  // (undocumented)
+  AppleTVPlay = "AppleTVPlay",
+  // (undocumented)
   Archive = "Archive",
   // (undocumented)
   AreaChart = "AreaChart",
@@ -3284,6 +3304,8 @@ enum IconNames {
   Attach = "Attach",
   // (undocumented)
   AustralianRules = "AustralianRules",
+  // (undocumented)
+  AuthenticatorApp = "AuthenticatorApp",
   // (undocumented)
   AutoEnhanceOff = "AutoEnhanceOff",
   // (undocumented)
@@ -3502,6 +3524,8 @@ enum IconNames {
   CalculatorMultiply = "CalculatorMultiply",
   // (undocumented)
   CalculatorNotEqualTo = "CalculatorNotEqualTo",
+  // (undocumented)
+  CalculatorPercentage = "CalculatorPercentage",
   // (undocumented)
   CalculatorSubtract = "CalculatorSubtract",
   // (undocumented)
@@ -3743,6 +3767,10 @@ enum IconNames {
   // (undocumented)
   ClearNight = "ClearNight",
   // (undocumented)
+  ClipboardList = "ClipboardList",
+  // (undocumented)
+  ClipboardListMirrored = "ClipboardListMirrored",
+  // (undocumented)
   ClipboardSolid = "ClipboardSolid",
   // (undocumented)
   Clock = "Clock",
@@ -3803,6 +3831,8 @@ enum IconNames {
   // (undocumented)
   ColumnVerticalSection = "ColumnVerticalSection",
   // (undocumented)
+  ColumnVerticalSectionEdit = "ColumnVerticalSectionEdit",
+  // (undocumented)
   Combine = "Combine",
   // (undocumented)
   Combobox = "Combobox",
@@ -3810,6 +3840,8 @@ enum IconNames {
   CommandPrompt = "CommandPrompt",
   // (undocumented)
   Comment = "Comment",
+  // (undocumented)
+  CommentActive = "CommentActive",
   // (undocumented)
   CommentAdd = "CommentAdd",
   // (undocumented)
@@ -3826,6 +3858,8 @@ enum IconNames {
   CompanyDirectory = "CompanyDirectory",
   // (undocumented)
   CompanyDirectoryMirrored = "CompanyDirectoryMirrored",
+  // (undocumented)
+  Compare = "Compare",
   // (undocumented)
   CompassNW = "CompassNW",
   // (undocumented)
@@ -3919,6 +3953,8 @@ enum IconNames {
   // (undocumented)
   Dataflows = "Dataflows",
   // (undocumented)
+  DataManagementSettings = "DataManagementSettings",
+  // (undocumented)
   DateTime = "DateTime",
   // (undocumented)
   DateTime2 = "DateTime2",
@@ -3946,6 +3982,8 @@ enum IconNames {
   DefaultRatio = "DefaultRatio",
   // (undocumented)
   DefectSolid = "DefectSolid",
+  // (undocumented)
+  DefenderTVM = "DefenderTVM",
   // (undocumented)
   Delete = "Delete",
   // (undocumented)
@@ -4247,6 +4285,8 @@ enum IconNames {
   // (undocumented)
   FabricFolder = "FabricFolder",
   // (undocumented)
+  FabricFolderConfirm = "FabricFolderConfirm",
+  // (undocumented)
   FabricFolderFill = "FabricFolderFill",
   // (undocumented)
   FabricFolderSearch = "FabricFolderSearch",
@@ -4483,6 +4523,8 @@ enum IconNames {
   // (undocumented)
   GenericScanFilled = "GenericScanFilled",
   // (undocumented)
+  GIF = "GIF",
+  // (undocumented)
   Giftbox = "Giftbox",
   // (undocumented)
   GiftboxOpen = "GiftboxOpen",
@@ -4595,6 +4637,8 @@ enum IconNames {
   // (undocumented)
   Hide2 = "Hide2",
   // (undocumented)
+  Hide3 = "Hide3",
+  // (undocumented)
   Highlight = "Highlight",
   // (undocumented)
   HighlightMappedShapes = "HighlightMappedShapes",
@@ -4610,6 +4654,8 @@ enum IconNames {
   HomeGroup = "HomeGroup",
   // (undocumented)
   HomeSolid = "HomeSolid",
+  // (undocumented)
+  HomeVerify = "HomeVerify",
   // (undocumented)
   HorizontalDistributeCenter = "HorizontalDistributeCenter",
   // (undocumented)
@@ -4645,6 +4691,8 @@ enum IconNames {
   // (undocumented)
   IncidentTriangle = "IncidentTriangle",
   // (undocumented)
+  IncomingCall = "IncomingCall",
+  // (undocumented)
   IncreaseIndentArrow = "IncreaseIndentArrow",
   // (undocumented)
   IncreaseIndentArrowMirrored = "IncreaseIndentArrowMirrored",
@@ -4659,7 +4707,11 @@ enum IconNames {
   // (undocumented)
   Info2 = "Info2",
   // (undocumented)
+  InformationBarriers = "InformationBarriers",
+  // (undocumented)
   InfoSolid = "InfoSolid",
+  // (undocumented)
+  InkingTool = "InkingTool",
   // (undocumented)
   InsertColumnsLeft = "InsertColumnsLeft",
   // (undocumented)
@@ -4674,6 +4726,8 @@ enum IconNames {
   InsertTextBox = "InsertTextBox",
   // (undocumented)
   InstallToDrive = "InstallToDrive",
+  // (undocumented)
+  InternalInvestigation = "InternalInvestigation",
   // (undocumented)
   InternetSharing = "InternetSharing",
   // (undocumented)
@@ -4843,6 +4897,8 @@ enum IconNames {
   // (undocumented)
   MapPinSolid = "MapPinSolid",
   // (undocumented)
+  MarkAsProtected = "MarkAsProtected",
+  // (undocumented)
   MarkDownLanguage = "MarkDownLanguage",
   // (undocumented)
   Market = "Market",
@@ -4854,6 +4910,8 @@ enum IconNames {
   MaximumValue = "MaximumValue",
   // (undocumented)
   Medal = "Medal",
+  // (undocumented)
+  MedalSolid = "MedalSolid",
   // (undocumented)
   Media = "Media",
   // (undocumented)
@@ -4882,6 +4940,12 @@ enum IconNames {
   MicrosoftFlowLogo = "MicrosoftFlowLogo",
   // (undocumented)
   MicrosoftStaffhubLogo = "MicrosoftStaffhubLogo",
+  // (undocumented)
+  MicrosoftTranslatorLogo = "MicrosoftTranslatorLogo",
+  // (undocumented)
+  MicrosoftTranslatorLogoBlue = "MicrosoftTranslatorLogoBlue",
+  // (undocumented)
+  MicrosoftTranslatorLogoGreen = "MicrosoftTranslatorLogoGreen",
   // (undocumented)
   MiniContract = "MiniContract",
   // (undocumented)
@@ -5069,6 +5133,8 @@ enum IconNames {
   // (undocumented)
   OpenFolderHorizontal = "OpenFolderHorizontal",
   // (undocumented)
+  OpenInNewTab = "OpenInNewTab",
+  // (undocumented)
   OpenInNewWindow = "OpenInNewWindow",
   // (undocumented)
   OpenPane = "OpenPane",
@@ -5166,6 +5232,8 @@ enum IconNames {
   PartyLeader = "PartyLeader",
   // (undocumented)
   PassiveAuthentication = "PassiveAuthentication",
+  // (undocumented)
+  PasswordField = "PasswordField",
   // (undocumented)
   Paste = "Paste",
   // (undocumented)
@@ -5525,6 +5593,8 @@ enum IconNames {
   // (undocumented)
   RepeatAll = "RepeatAll",
   // (undocumented)
+  RepeatOne = "RepeatOne",
+  // (undocumented)
   Reply = "Reply",
   // (undocumented)
   ReplyAll = "ReplyAll",
@@ -5551,7 +5621,11 @@ enum IconNames {
   // (undocumented)
   RepoSolid = "RepoSolid",
   // (undocumented)
+  ReservationOrders = "ReservationOrders",
+  // (undocumented)
   ResponsesMenu = "ResponsesMenu",
+  // (undocumented)
+  ReturnKey = "ReturnKey",
   // (undocumented)
   ReturnToSession = "ReturnToSession",
   // (undocumented)
@@ -5581,6 +5655,8 @@ enum IconNames {
   // (undocumented)
   RingerRemove = "RingerRemove",
   // (undocumented)
+  RingerSolid = "RingerSolid",
+  // (undocumented)
   Robot = "Robot",
   // (undocumented)
   Rocket = "Rocket",
@@ -5588,6 +5664,10 @@ enum IconNames {
   Room = "Room",
   // (undocumented)
   Rotate = "Rotate",
+  // (undocumented)
+  Rotate90Clockwise = "Rotate90Clockwise",
+  // (undocumented)
+  Rotate90CounterClockwise = "Rotate90CounterClockwise",
   // (undocumented)
   RowsChild = "RowsChild",
   // (undocumented)
@@ -5675,6 +5755,8 @@ enum IconNames {
   // (undocumented)
   Settings = "Settings",
   // (undocumented)
+  ShakeDevice = "ShakeDevice",
+  // (undocumented)
   Shapes = "Shapes",
   // (undocumented)
   Share = "Share",
@@ -5712,6 +5794,8 @@ enum IconNames {
   SidePanel = "SidePanel",
   // (undocumented)
   SidePanelMirrored = "SidePanelMirrored",
+  // (undocumented)
+  Signin = "Signin",
   // (undocumented)
   SignOut = "SignOut",
   // (undocumented)
@@ -5774,6 +5858,8 @@ enum IconNames {
   SliderThumb = "SliderThumb",
   // (undocumented)
   Slideshow = "Slideshow",
+  // (undocumented)
+  SmartGlassRemote = "SmartGlassRemote",
   // (undocumented)
   SnapToGrid = "SnapToGrid",
   // (undocumented)
@@ -6249,6 +6335,8 @@ enum IconNames {
   // (undocumented)
   Upload = "Upload",
   // (undocumented)
+  USB = "USB",
+  // (undocumented)
   UserEvent = "UserEvent",
   // (undocumented)
   UserFollowed = "UserFollowed",
@@ -6274,6 +6362,8 @@ enum IconNames {
   VennDiagram = "VennDiagram",
   // (undocumented)
   VerifiedBrand = "VerifiedBrand",
+  // (undocumented)
+  VerifiedBrandSolid = "VerifiedBrandSolid",
   // (undocumented)
   VersionControlPush = "VersionControlPush",
   // (undocumented)
@@ -6375,11 +6465,15 @@ enum IconNames {
   // (undocumented)
   WarningSolid = "WarningSolid",
   // (undocumented)
+  WavingHand = "WavingHand",
+  // (undocumented)
   WebComponents = "WebComponents",
   // (undocumented)
   WebPublish = "WebPublish",
   // (undocumented)
   Website = "Website",
+  // (undocumented)
+  WebTemplate = "WebTemplate",
   // (undocumented)
   Weights = "Weights",
   // (undocumented)
@@ -6808,6 +6902,12 @@ interface IDeclaredEventsByName {
 }
 
 // @public
+interface IDefaultSlotProps<TSlots> {
+  // (undocumented)
+  _defaultStyles: IComponentStyles<TSlots>;
+}
+
+// @public
 interface IDelayedRenderProps extends React.Props<{}> {
   delay?: number;
 }
@@ -6847,6 +6947,7 @@ interface IDetailsHeaderBaseProps extends React.ClassAttributes<DetailsHeaderBas
   ariaLabel?: string;
   ariaLabelForSelectAllCheckbox?: string;
   ariaLabelForSelectionColumn?: string;
+  ariaLabelForToggleAllGroupsButton?: string;
   className?: string;
   collapseAllVisibility?: CollapseAllVisibility;
   columnReorderOptions?: IColumnReorderOptions;
@@ -7138,7 +7239,7 @@ interface IDialogContentProps extends React.ClassAttributes<DialogContentBase> {
   subText?: string;
   subTextId?: string;
   theme?: ITheme;
-  title?: string;
+  title?: string | JSX.Element;
   titleId?: string;
   topButtonsProps?: IButtonProps[];
   type?: DialogType;
@@ -7239,7 +7340,7 @@ interface IDialogProps extends React.ClassAttributes<DialogBase>, IWithResponsiv
   subText?: string;
   theme?: ITheme;
   // @deprecated
-  title?: string;
+  title?: string | JSX.Element;
   // @deprecated
   topButtonsProps?: IButtonProps[];
   // @deprecated
@@ -7827,6 +7928,8 @@ interface IExtendedPersonaProps extends IPersonaProps {
   // (undocumented)
   isValid: boolean;
   // (undocumented)
+  key?: React.Key;
+  // (undocumented)
   shouldBlockSelection?: boolean;
 }
 
@@ -7920,6 +8023,11 @@ interface IFacepileStyles {
   root: IStyle;
   // (undocumented)
   screenReaderOnly: IStyle;
+}
+
+// @public
+interface IFactoryOptions<TProps> {
+  defaultProp?: keyof TProps | 'children';
 }
 
 // @public
@@ -8724,7 +8832,7 @@ interface IListState {
 }
 
 // @public (undocumented)
-class ImageBase extends BaseComponent<IImageProps, IImageState> {
+class ImageBase extends React.Component<IImageProps, IImageState> {
   constructor(props: IImageProps);
   // (undocumented)
   componentDidUpdate(prevProps: IImageProps, prevState: IImageState): void;
@@ -9683,6 +9791,12 @@ interface IPositioningContainerState {
   positions?: IPositionedData;
 }
 
+// @public
+interface IProcessedSlotProps {
+  // (undocumented)
+  className?: string;
+}
+
 // @public (undocumented)
 interface IProgressIndicatorProps extends React.ClassAttributes<ProgressIndicatorBase> {
   ariaValueText?: string;
@@ -10533,6 +10647,12 @@ interface ISliderStyles {
 }
 
 // @public
+interface ISlotCreator<TProps> {
+  // (undocumented)
+  create?: ISlotFactory<TProps>;
+}
+
+// @public
 export function isMac(reset?: boolean): boolean;
 
 // WARNING: Because this definition is explicitly marked as @internal, an underscore prefix ("_") should be added to its name
@@ -10733,6 +10853,18 @@ interface IStickyState {
   isStickyBottom: boolean;
   // (undocumented)
   isStickyTop: boolean;
+}
+
+// @public
+interface IStyleableComponentProps<TViewProps, TTokens, TStyleSet extends IStyleSet<TStyleSet>> {
+  // (undocumented)
+  className?: string;
+  // (undocumented)
+  styles?: IStylesFunctionOrObject<TViewProps, TTokens, TStyleSet>;
+  // (undocumented)
+  theme?: ITheme;
+  // (undocumented)
+  tokens?: ITokenFunctionOrObject<TViewProps, TTokens>;
 }
 
 // @public
@@ -11199,6 +11331,14 @@ interface ITheme extends IScheme {
 }
 
 // @public (undocumented)
+interface IThemeProviderProps {
+  // (undocumented)
+  scheme?: ISchemeNames;
+  // (undocumented)
+  theme?: ITheme;
+}
+
+// @public (undocumented)
 interface IThemeRules {
   // (undocumented)
   [key: string]: IThemeSlotRule;
@@ -11278,6 +11418,10 @@ interface IToggleStyles {
   root: IStyle;
   text: IStyle;
   thumb: IStyle;
+}
+
+// @public
+interface ITokenBaseArray<TViewProps, TTokens> extends Array<IToken<TViewProps, TTokens>> {
 }
 
 // @public (undocumented)
@@ -11513,7 +11657,7 @@ class ListPeoplePickerBase extends MemberListPeoplePicker {
   static defaultProps: {
     createGenericItem: typeof createGenericItem;
     onRenderItem: (props: IPeoplePickerItemSelectedProps) => JSX.Element;
-    onRenderSuggestionsItem: (personaProps: IPersonaProps, suggestionsProps?: IBasePickerSuggestionsProps | undefined) => JSX.Element;
+    onRenderSuggestionsItem: (personaProps: IPersonaProps, suggestionsProps?: IBasePickerSuggestionsProps<any> | undefined) => JSX.Element;
   }
 }
 
@@ -11651,7 +11795,7 @@ class NormalPeoplePickerBase extends BasePeoplePicker {
   static defaultProps: {
     createGenericItem: typeof createGenericItem;
     onRenderItem: (props: IPeoplePickerItemSelectedProps) => JSX.Element;
-    onRenderSuggestionsItem: (personaProps: IPersonaProps, suggestionsProps?: IBasePickerSuggestionsProps | undefined) => JSX.Element;
+    onRenderSuggestionsItem: (personaProps: IPersonaProps, suggestionsProps?: IBasePickerSuggestionsProps<any> | undefined) => JSX.Element;
   }
 }
 
@@ -12122,8 +12266,6 @@ class SelectedPeopleList extends BasePeopleSelectedItemsList {
   static defaultProps: any;
   // (undocumented)
   protected renderItems: () => JSX.Element[];
-  // (undocumented)
-  replaceItem: (itemToReplace: IExtendedPersonaProps, itemsToReplaceWith: IExtendedPersonaProps[]) => void;
 }
 
 // @public (undocumented)
@@ -12897,6 +13039,9 @@ export function warnDeprecations<P>(componentName: string, props: P, deprecation
 // @public
 export function warnMutuallyExclusive<P>(componentName: string, props: P, exclusiveMap: ISettingsMap<P>): void;
 
+// @public
+export function withSlots<P>(type: ISlot<P> | React.SFC<P> | string, props?: React.Attributes & P | null, ...children: React.ReactNode[]): React.ReactElement<P> | JSX.Element | null;
+
 // @public (undocumented)
 module ZIndexes {
   // (undocumented)
@@ -12978,6 +13123,34 @@ module ZIndexes {
 // WARNING: Unsupported export: IBaseFloatingPickerSuggestionProps
 // WARNING: Unsupported export: FocusZoneTabbableElements
 // WARNING: Unsupported export: FocusZoneTabbableElements
+// WARNING: Unsupported export: IPropsWithChildren
+// WARNING: Unsupported export: IComponentStyles
+// WARNING: Unsupported export: IStylesFunction
+// WARNING: Unsupported export: IStylesFunctionOrObject
+// WARNING: Unsupported export: IToken
+// WARNING: Unsupported export: ITokenFunction
+// WARNING: Unsupported export: ITokenFunctionOrObject
+// WARNING: Unsupported export: ITokenBase
+// WARNING: Unsupported export: ICustomizationProps
+// WARNING: Unsupported export: IStateComponentProps
+// WARNING: Unsupported export: IStateComponentType
+// WARNING: Unsupported export: IViewComponent
+// WARNING: Unsupported export: IViewRenderer
+// WARNING: Unsupported export: IHTMLSlot
+// WARNING: Unsupported export: IHTMLElementSlot
+// WARNING: Unsupported export: ISlottableComponentType
+// WARNING: Unsupported export: ISlottableReactType
+// WARNING: Unsupported export: ISlotDefinition
+// WARNING: Unsupported export: ISlot
+// WARNING: Unsupported export: ISlotFactory
+// WARNING: Unsupported export: ISlots
+// WARNING: Unsupported export: ISlotProps
+// WARNING: Unsupported export: ISlotProp
+// WARNING: Unsupported export: ISlotPropValue
+// WARNING: Unsupported export: ISlotPropRenderFunction
+// WARNING: Unsupported export: ISlotRenderer
+// WARNING: Unsupported export: ISlotRender
+// WARNING: Unsupported export: ThemeProvider
 // WARNING: Unsupported export: Grid
 // WARNING: Unsupported export: IGroupHeaderStyleProps
 // WARNING: Unsupported export: IGroupFooterStyleProps