+++ packages/office-ui-fabric-react/src/components/DetailsList/examples/DetailsList.CustomFooter.Example.tsx
@@ -1,5 +1,3 @@
-// @codepen
-
 import * as React from 'react';
 import {
   DetailsList,