+++ packages/office-ui-fabric-react/src/components/Dialog/DialogContent.types.ts
@@ -67,7 +67,7 @@ export interface IDialogContentProps extends React.ClassAttributes<DialogContent
   /**
    * The title text to display at the top of the dialog.
    */
-  title?: string;
+  title?: string | JSX.Element;
 
   /**
    * Responsive mode passed in from decorator.