+++ packages/office-ui-fabric-react/src/components/TextField/examples/TextField.CustomRender.Example.tsx
@@ -1,8 +1,14 @@
 import * as React from 'react';
 import { TextField, ITextFieldProps } from 'office-ui-fabric-react/lib/TextField';
-import { IconButton } from 'office-ui-fabric-react/lib/Button';
+import { DefaultButton, IconButton } from 'office-ui-fabric-react/lib/Button';
 import { Callout } from 'office-ui-fabric-react/lib/Callout';
+import { mergeStyles } from 'office-ui-fabric-react/lib/Styling';
+import { getId } from 'office-ui-fabric-react/lib/Utilities';
 import './TextField.Examples.scss';
+
+const calloutChildren = mergeStyles({
+  padding: '10px'
+});
 
 export class TextFieldCustomRenderExample extends React.Component<
   {},
@@ -11,6 +17,7 @@ export class TextFieldCustomRenderExample extends React.Component<
   }
 > {
   private _iconButtonElement: HTMLElement;
+  private _descriptionId: string = getId('description');
 
   constructor(props: {}) {
     super(props);
@@ -37,8 +44,20 @@ export class TextFieldCustomRenderExample extends React.Component<
           <IconButton iconProps={{ iconName: 'Info' }} title="Info" ariaLabel="Info" onClick={this._onClick} />
         </span>
         {isCalloutVisible && (
-          <Callout className="ms-CustomRenderExample-callout" target={this._iconButtonElement} onDismiss={this._onDismiss}>
-            <text> In additon to the label itself, this label includes an iconbutton which pops out more information in a callout</text>
+          <Callout
+            className="ms-CustomRenderExample-callout"
+            target={this._iconButtonElement}
+            setInitialFocus={true}
+            onDismiss={this._onDismiss}
+            ariaDescribedBy={this._descriptionId}
+            role="alertdialog"
+          >
+            <div className={calloutChildren} id={this._descriptionId}>
+              In addition to the label itself, this label includes an IconButton which pops out more information in a Callout
+            </div>
+            <div className={calloutChildren}>
+              <DefaultButton onClick={this._onDismiss}>Close</DefaultButton>
+            </div>
           </Callout>
         )}
       </div>