+++ apps/fabric-website/CHANGELOG.md
@@ -1,6 +1,27 @@
 # Change Log - @uifabric/fabric-website
 
-This log was last generated on Tue, 05 Mar 2019 04:25:07 GMT and should not be manually modified.
+This log was last generated on Fri, 22 Mar 2019 02:41:50 GMT and should not be manually modified.
+
+## 6.12.3
+Fri, 22 Mar 2019 02:41:50 GMT
+
+### Patches
+
+- Add missing hypen for prefixVer
+
+## 6.12.2
+Wed, 20 Mar 2019 12:34:45 GMT
+
+### Patches
+
+- Remove template literal in fabric-website/index
+
+## 6.12.1
+Tue, 19 Mar 2019 12:36:45 GMT
+
+### Patches
+
+- Add query parameter to change Fabric version
 
 ## 6.12.0
 Tue, 05 Mar 2019 04:25:07 GMT